user/pass register/verify works
still need to clean up and fix some stuff
