+++ src/db/models/index.ts
@@ -1,0 +1,6 @@
+import { User, UserModel } from "./user";
+
+export {
+  User,
+  UserModel,
+};
