+++ src/db/types/index.ts
@@ -1,0 +1,20 @@
+import { IDbUpdateInfo, IDbDeleteInfo } from "./db";
+import { OauthServiceType } from "./oauth";
+import {
+  UserAccountType,
+  IUser,
+  IUserSearchInfo,
+  IUserCreateInfo,
+  IUserUpdateInfo,
+} from "./user";
+
+export {
+  IDbUpdateInfo,
+  IDbDeleteInfo,
+  OauthServiceType,
+  UserAccountType,
+  IUser,
+  IUserSearchInfo,
+  IUserCreateInfo,
+  IUserUpdateInfo,
+};
