+++ src/types/env.d.ts
@@ -1,6 +1,7 @@
 declare namespace NodeJS {
   export interface ProcessEnv {
     NODE_ENV: "prod" | "dev" | "test";
+    NODE_PATH: string;
 
     WEBSERVER_LOGGER: string;
     WEBSERVER_PORT: string;
@@ -14,5 +15,9 @@ declare namespace NodeJS {
 
     EMAIL_DISPLAY: string;
     EMAIL_DISPLAY_NAME: string;
+
+    MONGO_HOST: string;
+    MONGO_DB: string;
+    MONGO_CERT: string;
   }
 }
